Refactored UseCaseInjector for each aggregate root.
