+++ core/src/main/java/com/aarves/bluepages/usecase/interactors/account/AccountUseCaseInjector.java
@@ -1,0 +1,19 @@
+package com.aarves.bluepages.usecase.interactors.account;
+
+import com.aarves.bluepages.usecase.data.AccountDAO;
+import com.aarves.bluepages.usecase.data.AccountDataAccess;
+
+import java.security.NoSuchAlgorithmException;
+
+public class AccountUseCaseInjector {
+    private final AccountManager accountManager;
+
+    public AccountUseCaseInjector(AccountDAO accountDAO, AccountOutputBoundary accountOutput) throws NoSuchAlgorithmException {
+        AccountDataBoundary accountDataAccess = new AccountDataAccess(accountDAO);
+        this.accountManager = new AccountManager(accountDataAccess, accountOutput);
+    }
+
+    public AccountManager getAccountManager() {
+        return this.accountManager;
+    }
+}
